+++ src/binomcikit/err/__init__.py
@@ -20,6 +20,7 @@ from .adj_all import (
 from .base_all import (
     errall,
     erras,
+    errblaker,
     errex,
     errlr,
     errlt,