+++ src/binomcikit/expl/__init__.py
@@ -21,6 +21,7 @@ from .base_all import (
     explall,
     lengthall,
     lengthas,
+    lengthblaker,
     lengthex,
     lengthlr,
     lengthlt,